chore: enable semantic release
