+++ package.json
@@ -19,7 +19,6 @@
     "build:debug": "npm run build -- --sourcemap",
     "typings": "tsc -b typings.tsconfig.json && mv lib/types.d.ts typings.d.ts && cat module-declaration.d.ts >> typings.d.ts",
     "prepublishOnly": "npm run build",
-    "publish-all": "node scripts/deploy.js",
     "size": "echo \"Gzipped, estimate: $(gzip -9 -c dist/cf-extension-api.js | wc -c) bytes\"",
     "integration": "ts-node test/integration",
     "integration:local": "ts-node test/integration/local",
@@ -27,12 +26,16 @@
     "cypress:run": "cypress run --browser chrome",
     "clean-reports": "rm -rf cypress/reports",
     "merge-report": "mochawesome-merge ./test/cypress/reports/separate-reports/*.json > ./test/cypress/reports/full_report.json",
-    "generate-report": "npx mochawesome-report-generator --reportDir test/cypress/reports test/cypress/reports/full_report.json"
+    "generate-report": "npx mochawesome-report-generator --reportDir test/cypress/reports test/cypress/reports/full_report.json",
+    "semantic-release": "semantic-release"
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "17.0.0",
     "@rollup/plugin-node-resolve": "11.0.1",
     "@rollup/plugin-typescript": "8.1.0",
+    "@semantic-release/changelog": "5.0.1",
+    "@semantic-release/exec": "5.0.0",
+    "@semantic-release/git": "9.0.0",
     "@testing-library/cypress": "7.0.3",
     "@testing-library/dom": "7.29.0",
     "@types/async-retry": "1.4.2",
@@ -75,6 +78,7 @@
     "prettier": "2.2.1",
     "rollup": "2.35.1",
     "rollup-plugin-terser": "7.0.2",
+    "semantic-release": "17.3.0",
     "sinon": "9.2.2",
     "sinon-chai": "3.5.0",
     "ts-mocha": "8.0.0",
@@ -95,6 +99,31 @@
     "*.md": [
       "prettier --write",
       "git add"
+    ]
+  },
+  "release": {
+    "plugins": [
+      "@semantic-release/commit-analyzer",
+      "@semantic-release/release-notes-generator",
+      "@semantic-release/changelog",
+      [
+        "@semantic-release/exec",
+        {
+          "verifyConditionsCmd": "node ./scripts/verify.js",
+          "publishCmd": "node ./scripts/publish.js"
+        }
+      ],
+      [
+        "@semantic-release/git",
+        {
+          "assets": [
+            "CHANGELOG.md",
+            "package.json",
+            "package-lock.json"
+          ]
+        }
+      ],
+      "@semantic-release/github"
     ]
   }
 }
